Split out RenameService. **TESTS NEED FIXING**
